General improvements for the system

+++ database/migrations/2024_04_20_132844_create_inventory_items_table.php
@@ -25,10 +25,14 @@ return new class extends Migration
             $table->string('barcode')->nullable();
             $table->longText('url_to_provider')->nullable();
             $table->longText('alt_url_to_provider')->nullable();
-            $table->bigInteger('total_count')->nullable();
-            $table->bigInteger('critical_amount')->nullable();
-            $table->bigInteger('to_order')->nullable();
-            $table->bigInteger('average_consumption')->nullable();
+            $table->double('total_count')->nullable();
+            $table->double('critical_amount')->nullable();
+            $table->double('to_order')->nullable();
+            $table->double('average_consumption')->nullable();
+//            $table->bigInteger('total_count')->nullable();
+//            $table->bigInteger('critical_amount')->nullable();
+//            $table->bigInteger('to_order')->nullable();
+//            $table->bigInteger('average_consumption')->nullable();
             $table->boolean('multiple_locations')->nullable();
             $table->foreignId('laboratory')->nullable()->constrained('laboratories');
             $table->foreignId('cupboard')->nullable()->constrained('cupboards');
